# FinancialApp-main/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /ledger/transfer:
    post:
      operationId: LedgerController_createTransfer
      summary: Создать перевод через главную книгу
      parameters:
        - name: idempotency-key
          required: true
          in: header
          schema:
            type: string
        - name: x-trace-id
          required: true
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerEntryDto"
      responses:
        "201":
          description: Перевод успешно создан
        "400":
          description: Неверные параметры запроса
        "409":
          description: Конфликт версий (нужно повторить запрос)
      tags: &a5
        - ledger
      security: &a6
        - bearer: []
  /ledger/account/{accountId}:
    get:
      operationId: LedgerController_findByAccount
      summary: Получить историю транзакций для счета
      parameters:
        - name: accountId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: История транзакций
      tags: *a5
      security: *a6
  /ledger/{accountId}:
    get:
      operationId: LedgerController_findByAccountLegacy
      summary: Получить историю транзакций для счета (legacy path)
      parameters:
        - name: accountId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: История транзакций
      tags: *a5
      security: *a6
  /ledger/entry/{entryId}:
    get:
      operationId: LedgerController_getEntry
      summary: Получить запись по ID
      parameters:
        - name: entryId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: Запись найдена
        "404":
          description: Запись не найдена
      tags: *a5
      security: *a6
  /ledger/account/{accountId}/balance/verify:
    get:
      operationId: LedgerController_verifyBalance
      summary: Проверить целостность баланса счета
      parameters:
        - name: accountId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: Результат проверки баланса
      tags: *a5
      security: *a6
  /ledger/account/{accountId}/reconcile:
    post:
      operationId: LedgerController_reconcileAccount
      summary: Реконсиляция баланса счета с проводками
      parameters:
        - name: accountId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: Результат реконсиляции
      tags: *a5
      security: *a6
...
  /payments:
    post:
      operationId: PaymentsController_create
      parameters:
        - name: idempotency-key
          required: true
          in: header
          schema:
            type: string
        - name: x-trace-id
          required: true
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentDto"
      responses:
        "201":
          description: ""
      tags: &a9
        - payments
      security: &a10
        - bearer: []
  /payments/schedules:
    get:
      operationId: PaymentsController_listSchedules
      parameters: []
      responses:
        "200":
          description: ""
      tags: *a9
      security: *a10
  /payments/schedules/{id}/cancel:
    post:
      operationId: PaymentsController_cancelSchedule
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "201":
          description: ""
      tags: *a9
      security: *a10
...
  /notifications:
    get:
      operationId: NotificationsController_getNotifications
      summary: Get user notifications with pagination
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - name: type
          required: false
          in: query
          schema:
            enum:
              - TRANSACTION
              - PAYMENT
              - CARD_CONTROL
              - KYC
              - LIMIT
              - SYSTEM
            type: string
        - name: unreadOnly
          required: false
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
      tags: &a15
        - notifications
      security: &a16
        - bearer: []
  /notifications/unread/count:
    get:
      operationId: NotificationsController_getUnreadCount
      summary: Get count of unread notifications
      parameters: []
      responses:
        "200":
          description: Count of unread notifications
      tags: *a15
      security: *a16
  /notifications/devices:
    post:
      operationId: NotificationsController_registerDevice
      summary: Register a push notification device token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceDto"
      responses:
        "201":
          description: Device registered
      tags: *a15
      security: *a16
  /notifications/{id}:
    get:
      operationId: NotificationsController_getNotification
      summary: Get notification by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: Notification found
        "404":
          description: Notification not found
      tags: *a15
      security: *a16
  /notifications/{id}/read:
    patch:
      operationId: NotificationsController_markAsRead
      summary: Mark notification as read
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: Notification marked as read
        "404":
          description: Notification not found
      tags: *a15
      security: *a16
  /notifications/read-all:
    patch:
      operationId: NotificationsController_markAllAsRead
      summary: Mark all notifications as read
      parameters: []
      responses:
        "200":
          description: All notifications marked as read
      tags: *a15
      security: *a16
  /notifications/preferences:
    post:
      operationId: NotificationsController_update
      summary: Update notification preferences
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationsDto"
      responses:
        "200":
          description: Preferences updated
      tags: *a15
      security: *a16
    get:
      operationId: NotificationsController_get
      summary: Get notification preferences
      parameters: []
      responses:
        "200":
          description: User notification preferences
      tags: *a15
      security: *a16
...
info:
  title: Digital Banking Core API
  description: API surface for the digital banking MVP core components.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateLedgerEntryDto:
      type: object
      properties:
        debitAccountId:
          type: number
        creditAccountId:
          type: number
        amount:
          type: number
        currency:
          type: string
          default: USD
        idempotencyKey:
          type: string
          description: Idempotency key per client request
        traceId:
          type: string
      required:
        - debitAccountId
        - creditAccountId
        - amount
        - currency
        - idempotencyKey
    CreatePaymentDto:
      type: object
      properties:
        fromAccount:
          type: number
        toAccount:
          type: number
        amount:
          type: number
        currency:
          type: string
          default: USD
        idempotencyKey:
          type: string
        traceId:
          type: string
        transferType:
          type: string
          enum:
            - INTERNAL
            - INTERBANK
          default: INTERNAL
        beneficiaryIban:
          type: string
          description: Beneficiary IBAN for interbank transfers
        beneficiaryBank:
          type: string
          description: Beneficiary bank name
        scheduledFor:
          type: string
          description: Schedule payment for a future ISO date
        description:
          type: string
        cardToken:
          type: string
          description: Card token for card transactions
        mcc:
          type: number
          description: MCC (Merchant Category Code) for card transactions
        geoLocation:
          type: string
          description: Geolocation code (ISO country code) for card transactions
          example: US
      required:
        - fromAccount
        - toAccount
        - amount
        - currency
        - idempotencyKey
...
    RegisterDeviceDto:
      type: object
      properties: {}
    UpdateNotificationsDto:
      type: object
      properties: {}
...